feat(ui): self-serve ui for pearl orbs
based on diamond ui
events for Pearl are forwarded to a different
handler depending on mode of operation
