+++ ui/rgb/src/lib.rs
@@ -22,21 +22,37 @@ impl ops::Mul<f64> for Argb {
 
     #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
     fn mul(self, rhs: f64) -> Self::Output {
-        Argb(
-            self.0,
-            ((f64::from(self.1) * rhs) as u8).clamp(0, u8::MAX),
-            ((f64::from(self.2) * rhs) as u8).clamp(0, u8::MAX),
-            ((f64::from(self.3) * rhs) as u8).clamp(0, u8::MAX),
-        )
+        // at low brightness (low rhs)
+        // prefer to turn LED off if the resulting color has only 1 component
+        // with an initial color that has more than 1 component
+        if ((self.1 != 0) as u8) + ((self.2 != 0) as u8) + ((self.3 != 0) as u8) > 1 {
+            let res = Argb(
+                self.0,
+                ((f64::from(self.1) * rhs) as u8).clamp(0, u8::MAX),
+                ((f64::from(self.2) * rhs) as u8).clamp(0, u8::MAX),
+                ((f64::from(self.3) * rhs) as u8).clamp(0, u8::MAX),
+            );
+            // if result color has only 1 component, prefer to turn LED off
+            if ((res.1 != 0) as u8) + ((res.2 != 0) as u8) + ((res.3 != 0) as u8) == 1 {
+                Argb::OFF
+            } else {
+                res
+            }
+        } else {
+            Argb(
+                self.0,
+                ((f64::from(self.1) * rhs) as u8).clamp(0, u8::MAX),
+                ((f64::from(self.2) * rhs) as u8).clamp(0, u8::MAX),
+                ((f64::from(self.3) * rhs) as u8).clamp(0, u8::MAX),
+            )
+        }
     }
 }
 
 impl ops::MulAssign<f64> for Argb {
     #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
     fn mul_assign(&mut self, rhs: f64) {
-        self.1 = ((f64::from(self.1) * rhs) as u8).clamp(0, u8::MAX);
-        self.2 = ((f64::from(self.2) * rhs) as u8).clamp(0, u8::MAX);
-        self.3 = ((f64::from(self.3) * rhs) as u8).clamp(0, u8::MAX);
+        *self = *self * rhs;
     }
 }
 
@@ -69,6 +85,29 @@ impl Argb {
     pub const PEARL_USER_SIGNUP: Argb = Argb(None, 31, 31, 31);
     pub const PEARL_USER_FLASH: Argb = Argb(None, 255, 255, 255);
 
+    /// ***** Self-serve colors *****
+    /// We intentionally don't include blue in most of the color scheme
+    /// because a sine wave with a low blue component doesn't look good:
+    /// whiter once wave is over, but darker during the wave.
+    ///
+    /// Outer-ring color during operator QR scans
+    pub const PEARL_RING_OPERATOR_QR_SCAN: Argb = Argb(None, 20, 6, 0);
+    pub const PEARL_RING_OPERATOR_QR_SCAN_SPINNER: Argb = Argb(None, 25, 15, 9);
+    pub const PEARL_RING_OPERATOR_QR_SCAN_SPINNER_OPERATOR_BASED: Argb =
+        Argb(None, 25, 22, 5);
+    /// Outer-ring color during user QR scans
+    pub const PEARL_RING_USER_QR_SCAN: Argb = Argb(None, 30, 20, 0);
+    pub const PEARL_RING_USER_QR_SCAN_SPINNER: Argb = Argb(None, 28, 25, 10);
+    /// Shroud color to invite user to scan / reposition in front of the orb
+    pub const PEARL_CENTER_SUMMON_USER_AMBER: Argb = Argb(None, 30, 20, 0);
+    /// Shroud color during user scan/capture (in progress)
+    pub const PEARL_CENTER_USER_CAPTURE: Argb = Argb(None, 30, 20, 0);
+    /// Outer-ring color during user scan/capture (in progress)
+    pub const PEARL_RING_USER_CAPTURE: Argb = Argb(None, 30, 20, 0);
+    /// Error color for outer ring
+    pub const PEARL_RING_ERROR_SALMON: Argb = Argb(None, 24, 4, 0);
+
+    /// ***** Self-serve colors *****
     pub const DIAMOND_OPERATOR_AMBER: Argb =
         Argb(Some(Self::DIMMING_MAX_VALUE), 20, 16, 0);
     // To help quickly distinguish dev vs prod software,
@@ -87,10 +126,8 @@ impl Argb {
     /// Outer-ring color during user QR scans
     pub const DIAMOND_RING_USER_QR_SCAN: Argb = Argb(Some(5), 120, 80, 4);
     pub const DIAMOND_RING_USER_QR_SCAN_SPINNER: Argb = Argb(Some(10), 100, 90, 35);
-    /// Shroud color to invite user to scan / reposition in front of the orb
-    pub const DIAMOND_SHROUD_SUMMON_USER_AMBER: Argb = Argb(Some(3), 95, 40, 3);
-    /// Shroud color during user scan/capture (in progress)
-    pub const DIAMOND_SHROUD_USER_CAPTURE: Argb = Argb(Some(3), 118, 51, 3);
+    /// Shroud color to invite user to scan / reposition in front of the orb and capture
+    pub const DIAMOND_CENTER_SUMMON_USER_AMBER: Argb = Argb(Some(3), 95, 40, 3);
     /// Outer-ring color during user scan/capture (in progress)
     pub const DIAMOND_RING_USER_CAPTURE: Argb = Argb(Some(10), 120, 100, 4);
     pub const DIAMOND_CONE_AMBER: Argb = Argb(Some(Self::DIMMING_MAX_VALUE), 25, 18, 1);
